move header into its own component
