+++ assets/js/root.jsx
@@ -4,6 +4,7 @@ import { Link, BrowserRouter as Router, Route } from 'react-router-dom';
 import _ from 'lodash';
 import $ from 'jquery';
 
+import Header from './Header';
 import Notification from './Notification';
 import Post from './Post';
 import UserProfile from './UserProfile';
@@ -149,50 +150,6 @@ class Root extends React.Component {
     }
 }
 
-function Header(props) {
-    let {root, session} = props;
-    let session_info;
-    if (session == null) {
-        session_info = <div className="form-inline my-2">
-            <form action="javascript:void(0)" onSubmit={() => root.login()}>
-            <input type="email" placeholder="email"
-        onChange={(ev) => root.update_login_form({email: ev.target.value})} />
-            <input type="password" placeholder="password"
-        onChange={(ev) => root.update_login_form({password: ev.target.value})} />
-            <button className="btn btn-secondary" type="submit">Login</button>
-            <br/>
-            {root.state.error || ""}
-            </form>
-            </div>;
-    } else {
-        session_info = <div className="my-2">
-            <p><font color="yellow">Logged in as </font>{session.display_name}</p>
-            <Link to={"/edituser"}>Edit Profile</Link><br/>
-            <button className="btn btn-secondary" onClick={() => root.logout()}>Log Out</button>
-            </div>
-    }
-
-    let register_link = [];
-    if (session == null) {
-        register_link = <Link to={"/register"}>Register</Link>;
-    }
-
-    return <div className="row my-2">
-        <div className="col-4">
-            <Link to={"/"}><h1>Pokester</h1></Link>
-        </div>
-        <div className="col-4">
-        <p>
-        <Link to={"/users"}>Users</Link>
-        &nbsp;
-        {register_link}
-        </p>
-        </div>
-        <div className="col-4">
-        {session_info}
-        </div>
-        </div>;
-}
 
 function UserList(props) {
       let rows = _.map(props.users, (uu) => <User key={uu.id} user={uu} />);
